refactor(console): active custom domain process
