+++ packages/console/package.json
@@ -61,7 +61,6 @@
     "@types/react-modal": "^3.13.1",
     "@types/react-syntax-highlighter": "^15.5.1",
     "buffer": "^5.7.1",
-    "camelcase": "^7.0.0",
     "classnames": "^2.3.1",
     "clean-deep": "^3.4.0",
     "cross-env": "^7.0.3",
