+++ packages/console/src/pages/TenantSettings/tabs/TenantDomainSettings/components/CustomDomain/components/ActivationProcess/components/DnsRecordsTable/index.module.scss
