+++ packages/console/src/pages/TenantSettings/tabs/TenantDomainSettings/components/CustomDomain/index.tsx
@@ -1,6 +1,6 @@
 import { type Domain, DomainStatus } from '@logto/schemas';
 
-import ActiveProcess from './components/ActiveProcess';
+import ActivationProcess from './components/ActivationProcess';
 import CustomDomainHeader from './components/CustomDomainHeader';
 import * as styles from './index.module.scss';
 
@@ -13,7 +13,9 @@ function CustomDomain({ customDomain, onDeleteCustomDomain }: Props) {
   return (
     <div className={styles.container}>
       <CustomDomainHeader customDomain={customDomain} onDeleteCustomDomain={onDeleteCustomDomain} />
-      {customDomain.status !== DomainStatus.Active && <ActiveProcess customDomain={customDomain} />}
+      {customDomain.status !== DomainStatus.Active && (
+        <ActivationProcess customDomain={customDomain} />
+      )}
     </div>
   );
 }
